moved plugin validation to separated file
